requirements.txt
gdown
pandas
Pillow
torch
tqdm
matplotlib
scikit-learn
torchvision